extract common logic
