+++ src/frontend/src/optimizer/plan_node/derive.rs
@@ -1,0 +1,104 @@
+use std::collections::HashSet;
+
+use fixedbitset::FixedBitSet;
+use itertools::Itertools;
+use risingwave_common::catalog::{ColumnCatalog, ColumnDesc, Schema, USER_COLUMN_ID_OFFSET};
+use risingwave_common::error::{ErrorCode, Result};
+
+use super::PlanRef;
+use crate::optimizer::property::{Direction, FieldOrder, Order};
+
+pub(crate) fn derive_columns(
+    schema: &Schema,
+    out_names: Vec<String>,
+    user_cols: &FixedBitSet,
+) -> Result<Vec<ColumnCatalog>> {
+    // Validate and deduplicate column names.
+    let mut col_names = HashSet::new();
+    for name in &out_names {
+        if !col_names.insert(name.clone()) {
+            Err(ErrorCode::InvalidInputSyntax(format!(
+                "column \"{}\" specified more than once",
+                name
+            )))?;
+        }
+    }
+
+    let mut out_name_iter = out_names.into_iter();
+    let columns = schema
+        .fields()
+        .iter()
+        .enumerate()
+        .map(|(i, field)| {
+            let mut c = ColumnCatalog {
+                column_desc: ColumnDesc::from_field_with_column_id(
+                    field,
+                    i as i32 + USER_COLUMN_ID_OFFSET,
+                ),
+                is_hidden: !user_cols.contains(i),
+            };
+            c.column_desc.name = if !c.is_hidden {
+                out_name_iter.next().unwrap()
+            } else {
+                let mut name = field.name.clone();
+                let mut count = 0;
+
+                while !col_names.insert(name.clone()) {
+                    count += 1;
+                    name = format!("{}#{}", field.name, count);
+                }
+
+                name
+            };
+            c
+        })
+        .collect_vec();
+
+    Ok(columns)
+}
+
+/// Derive the pk and the stream key for tables and sinks.
+pub(crate) fn derive_pk(
+    input: PlanRef,
+    user_order_by: Order,
+    columns: &[ColumnCatalog],
+) -> (Vec<FieldOrder>, Vec<usize>) {
+    // Note(congyi): avoid pk duplication
+    let stream_key = input.logical_pk().iter().copied().unique().collect_vec();
+    let schema = input.schema();
+
+    // Assert the uniqueness of column names and IDs, including hidden columns.
+    if let Some(name) = columns.iter().map(|c| c.name()).duplicates().next() {
+        panic!("duplicated column name \"{name}\"");
+    }
+    if let Some(id) = columns.iter().map(|c| c.column_id()).duplicates().next() {
+        panic!("duplicated column ID {id}");
+    }
+    // Assert that the schema of given `columns` is correct.
+    assert_eq!(
+        columns.iter().map(|c| c.data_type().clone()).collect_vec(),
+        input.schema().data_types()
+    );
+
+    let mut in_order = FixedBitSet::with_capacity(schema.len());
+    let mut pk = vec![];
+
+    for field in &user_order_by.field_order {
+        let idx = field.index;
+        pk.push(field.clone());
+        in_order.insert(idx);
+    }
+
+    for &idx in &stream_key {
+        if in_order.contains(idx) {
+            continue;
+        }
+        pk.push(FieldOrder {
+            index: idx,
+            direct: Direction::Asc,
+        });
+        in_order.insert(idx);
+    }
+
+    (pk, stream_key)
+}
