+++ src/frontend/src/optimizer/plan_node/stream_materialize.rs
@@ -13,20 +13,21 @@
 // limitations under the License.
 
 use std::assert_matches::assert_matches;
-use std::collections::HashSet;
 use std::fmt;
 
 use fixedbitset::FixedBitSet;
 use itertools::Itertools;
-use risingwave_common::catalog::{ColumnCatalog, ColumnDesc, TableId, USER_COLUMN_ID_OFFSET};
-use risingwave_common::error::{ErrorCode, Result};
+use risingwave_common::catalog::{ColumnCatalog, TableId};
+use risingwave_common::error::Result;
 use risingwave_pb::stream_plan::stream_node::NodeBody as ProstStreamNode;
 
+use super::derive::derive_columns;
 use super::{ExprRewritable, PlanRef, PlanTreeNodeUnary, StreamNode};
 use crate::catalog::table_catalog::{TableCatalog, TableType, TableVersion};
 use crate::catalog::FragmentId;
+use crate::optimizer::plan_node::derive::derive_pk;
 use crate::optimizer::plan_node::{PlanBase, PlanNode};
-use crate::optimizer::property::{Direction, Distribution, FieldOrder, Order, RequiredDist};
+use crate::optimizer::property::{Distribution, FieldOrder, Order, RequiredDist};
 use crate::stream_fragmenter::BuildFragmentGraphState;
 
 /// Materializes a stream.
@@ -61,48 +62,7 @@ impl StreamMaterialize {
         table_type: TableType,
     ) -> Result<Self> {
         let input = Self::rewrite_input(input, user_distributed_by, table_type)?;
-        let schema = input.schema();
-
-        // Used to validate and deduplicate column names.
-        let mut col_names = HashSet::new();
-        for name in &out_names {
-            if !col_names.insert(name.clone()) {
-                Err(ErrorCode::InvalidInputSyntax(format!(
-                    "column \"{}\" specified more than once",
-                    name
-                )))?;
-            }
-        }
-        let mut out_name_iter = out_names.into_iter();
-
-        let columns = schema
-            .fields()
-            .iter()
-            .enumerate()
-            .map(|(i, field)| {
-                let mut c = ColumnCatalog {
-                    column_desc: ColumnDesc::from_field_with_column_id(
-                        field,
-                        i as i32 + USER_COLUMN_ID_OFFSET,
-                    ),
-                    is_hidden: !user_cols.contains(i),
-                };
-                c.column_desc.name = if !c.is_hidden {
-                    out_name_iter.next().unwrap()
-                } else {
-                    let mut name = field.name.clone();
-                    let mut count = 0;
-
-                    while !col_names.insert(name.clone()) {
-                        count += 1;
-                        name = format!("{}#{}", field.name, count);
-                    }
-
-                    name
-                };
-                c
-            })
-            .collect_vec();
+        let columns = derive_columns(input.schema(), out_names, &user_cols)?;
 
         let table = Self::derive_table_catalog(
             input.clone(),
@@ -199,60 +159,25 @@ impl StreamMaterialize {
     ) -> Result<TableCatalog> {
         let input = rewritten_input;
 
-        let watermark_columns = input.watermark_columns().clone();
-        // Note(congyi): avoid pk duplication
-        let pk_indices = input.logical_pk().iter().copied().unique().collect_vec();
-        let schema = input.schema();
-        let distribution = input.distribution();
-
-        // Assert the uniqueness of column names and IDs, including hidden columns.
-        if let Some(name) = columns.iter().map(|c| c.name()).duplicates().next() {
-            panic!("duplicated column name \"{name}\"");
-        }
-        if let Some(id) = columns.iter().map(|c| c.column_id()).duplicates().next() {
-            panic!("duplicated column ID {id}");
-        }
-        // Assert that the schema of given `columns` is correct.
-        assert_eq!(
-            columns.iter().map(|c| c.data_type().clone()).collect_vec(),
-            input.schema().data_types()
-        );
-
         let value_indices = (0..columns.len()).collect_vec();
-        let mut in_order = FixedBitSet::with_capacity(schema.len());
-        let mut pk_list = vec![];
-
-        for field in &user_order_by.field_order {
-            let idx = field.index;
-            pk_list.push(field.clone());
-            in_order.insert(idx);
-        }
-
-        for &idx in &pk_indices {
-            if in_order.contains(idx) {
-                continue;
-            }
-            pk_list.push(FieldOrder {
-                index: idx,
-                direct: Direction::Asc,
-            });
-            in_order.insert(idx);
-        }
-
-        let distribution_key = distribution.dist_column_indices().to_vec();
+        let distribution_key = input.distribution().dist_column_indices().to_vec();
         let properties = input.ctx().with_options().internal_table_subset(); // TODO: remove this
-        let read_prefix_len_hint = pk_indices.len();
+        let append_only = input.append_only();
+        let watermark_columns = input.watermark_columns().clone();
+
+        let (pk, stream_key) = derive_pk(input, user_order_by, &columns);
+        let read_prefix_len_hint = stream_key.len();
 
         Ok(TableCatalog {
             id: TableId::placeholder(),
             associated_source_id: None,
             name,
             columns,
-            pk: pk_list,
-            stream_key: pk_indices,
+            pk,
+            stream_key,
             distribution_key,
             table_type,
-            append_only: input.append_only(),
+            append_only,
             owner: risingwave_common::catalog::DEFAULT_SUPER_USER_ID,
             properties,
             // TODO(zehua): replace it with FragmentId::placeholder()
